fix: 统一状态和枚举值的 int16 类型适配
本次提交统一了多个模块中状态、枚举值的类型，将原本使用 int32 的常量和字段修改为 int16，同时调整了数据库配置路径和 MySQL 迁移 PostgreSQL 的数据源配置，适配类型变更的同时修复了配置路径错误和数据库类型不匹配问题。

+++ backend/adapter/kit/redact_policy_resolver.go
@@ -128,7 +128,7 @@ func (r *RedactPolicyResolver) refreshLocked(ctx context.Context) error {
 
 	storageQuery := r.storagePolicyRepository.Query(ctx).BaseRedactStoragePolicy
 	storageOpts := make([]repository.QueryOption, 0, 2)
-	storageOpts = append(storageOpts, repository.Where(storageQuery.Status.Eq(_const.STATUS_STATUS_ENABLE)))
+	storageOpts = append(storageOpts, repository.Where(storageQuery.Status.Eq(int16(_const.STATUS_STATUS_ENABLE))))
 	storageOpts = append(storageOpts, repository.Order(storageQuery.ID.Asc()))
 	var storageRows []*models.BaseRedactStoragePolicy
 	storageRows, err = r.storagePolicyRepository.List(ctx, storageOpts...)
@@ -143,7 +143,7 @@ func (r *RedactPolicyResolver) refreshLocked(ctx context.Context) error {
 
 	outputQuery := r.outputPolicyRepository.Query(ctx).BaseRedactOutputPolicy
 	outputOpts := make([]repository.QueryOption, 0, 2)
-	outputOpts = append(outputOpts, repository.Where(outputQuery.Status.Eq(_const.STATUS_STATUS_ENABLE)))
+	outputOpts = append(outputOpts, repository.Where(outputQuery.Status.Eq(int16(_const.STATUS_STATUS_ENABLE))))
 	outputOpts = append(outputOpts, repository.Order(outputQuery.ID.Asc()))
 	var outputRows []*models.BaseRedactOutputPolicy
 	outputRows, err = r.outputPolicyRepository.List(ctx, outputOpts...)
